+++ common/include/common/config/GameConfig.h
@@ -39,7 +39,7 @@ struct GameConfig
 
     // Graphics
     CfgVar<bool> anisotropic_filtering = true;
-    CfgVar<unsigned> msaa = 0;
+    CfgVar<uint32_t> msaa_level = 0;
 
     CfgVar<bool> high_scanner_res = true;
     CfgVar<bool> true_color_textures = true;