+++ game_patch/graphics/d3d11/gr_d3d11_hooks.h
@@ -2,7 +2,7 @@
 
 #include "../../rf/gr/gr.h"
 
-namespace df::gr::d3d11 {
+namespace gr::d3d11 {
     bool set_render_target(int bm_handle);
     void update_window_mode();
     void bitmap_float(int bitmap_handle, float x, float y, float w, float h,