+++ game_patch/rf/bmpman.h
@@ -100,4 +100,5 @@ namespace rf::bm
     static auto& bm_set_resolution_level_for_all = addr_as_ref<void(int level)>(0x0050EF10);
 
     static auto& bitmaps = addr_as_ref<BitmapEntry*>(0x017C80C4);
+    static auto& max_bitmaps = addr_as_ref<int>(0x005A4554);
 }